requirements.txt
astroid==2.0.4
autopep8==1.4.2
backcall==0.1.0
bleach==3.0.2
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.4
opencv-contrib-python==3.4.4.19
opencv-python==3.4.4.19
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycodestyle==2.4.0
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
pyzmq==17.1.2
Send2Trash==1.5.0
six==1.11.0
terminado==0.8.1
testpath==0.4.2
torch==1.0.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
typed-ast==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.11
